[web] Hide JS types from dart:ui_web (#42252)
In order to use the exported `PlatformLocation` cross-platform, we need to remove any JS types from the interface. Namely, `DomEventListener`.

Required by https://github.com/flutter/flutter/pull/126851

+++ lib/web_ui/lib/src/engine/dom.dart
@@ -2119,7 +2119,7 @@ extension DomHistoryExtension on DomHistory {
   external JSVoid _go1();
   @JS('go')
   external JSVoid _go2(JSNumber delta);
-  void go([double? delta]) {
+  void go([int? delta]) {
     if (delta == null) {
       _go1();
     } else {
